% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsa_2018.R
\docType{data}
\name{cbsa_2018}
\alias{cbsa_2018}
\title{Core-Based Statistical Areas 2018 rural definition}
\format{
A tibble with 5 variables:
\describe{
  \item{geoid}{string County GEOID code}
  \item{name}{string Name of the rural definition}
  \item{year}{dbl Year the definition was released}
  \item{rural_def}{string Detailed rural classification: Metro, Micro, or Non-CBSA}
  \item{is_rural}{string Value of either Rural or Nonrural}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2018.html#list-tab-790442341}
}
\usage{
cbsa_2018
}
\description{
County-level rural classification based on the 2018 OMB mid-decade
delineation update.
}
\keyword{datasets}
