% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsa_2023.R
\docType{data}
\name{cbsa_2023}
\alias{cbsa_2023}
\title{Core-Based Statistical Areas 2023 rural definition}
\format{
A tibble with 3235 rows and 5 variables:
\describe{
  \item{geoid}{string County GEOID code}
  \item{name}{string Name of the rural definition}
  \item{year}{dbl Year the definition was released}
  \item{rural_def}{string Detailed rural classification if relevant}
  \item{is_rural}{string Value of either Rural or Nonrural}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.2023.html#list-tab-790442341}
}
\usage{
cbsa_2023
}
\description{
County-level rural classification
}
\keyword{datasets}
