% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchs_2013.R
\docType{data}
\name{nchs_2013}
\alias{nchs_2013}
\title{National Center for Health Statistics 2013 rural definition}
\format{
A tibble with 3133 rows and 5 variables:
\describe{
  \item{geoid}{string County GEOID code}
  \item{name}{string Name of the rural definition}
  \item{year}{dbl Year the definition was released}
  \item{rural_def}{string Detailed rural classification if relevant}
  \item{is_rural}{string Value of either Rural or Nonrural}
}
}
\source{
\url{https://www.cdc.gov/nchs/data-analysis-tools/urban-rural.html}
}
\usage{
nchs_2013
}
\description{
County-level rural classification
}
\keyword{datasets}
