% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rucc_2003.R
\docType{data}
\name{rucc_2003}
\alias{rucc_2003}
\title{2003 Rural Urban Continuum Codes}
\format{
A tibble with 3142 rows and 5 variables:
\describe{
  \item{geoid}{string County GEOID code}
  \item{name}{string Name of the rural definition}
  \item{year}{dbl Year the definition was released}
  \item{rural_def}{string Detailed rural classification if relevant}
  \item{is_rural}{string Value of either Rural or Nonrural}
}
}
\source{
\url{https://www.ers.usda.gov/data-products/rural-urban-continuum-codes}
}
\usage{
rucc_2003
}
\description{
County-level rural classification
}
\keyword{datasets}
