% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruraldefinitions-package.R
\docType{package}
\name{ruraldefinitions-package}
\alias{ruraldefinitions}
\alias{ruraldefinitions-package}
\title{ruraldefinitions: Data package for accessing and manipulating rural definitions}
\description{
This package aggregates common rural definitions and makes them accessible to R programmers. County-level rural definitions are included within the package, while rural definitions at more detailed geographic levels (such as Census tract) are loaded from an S3 bucket.
}
\author{
\strong{Maintainer}: Camden Blatchly \email{camden.blatchly@ruralinnovation.us}

}
\keyword{internal}
