% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uic_1993.R
\docType{data}
\name{uic_1993}
\alias{uic_1993}
\title{1993 Urban Influence Codes}
\format{
A tibble with 3233 rows and 5 variables:
\describe{
  \item{geoid}{string County GEOID code}
  \item{name}{string Name of the rural definition}
  \item{year}{dbl Year the definition was released}
  \item{rural_def}{string Detailed rural classification if relevant}
  \item{is_rural}{string Value of either Rural or Nonrural}
}
}
\source{
\url{https://www.ers.usda.gov/data-products/urban-influence-codes}
}
\usage{
uic_1993
}
\description{
County-level rural classification
}
\keyword{datasets}
