% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definition.R
\name{get_definition}
\alias{get_definition}
\title{Get definition from S3}
\usage{
get_definition(name, year)
}
\arguments{
\item{name}{Name of the rural definition}

\item{year}{Publication year of the desired definition}
}
\description{
Get definition from S3
}
